Add new activity to show projects

+++ app/src/main/java/com/github/mobile/api/service/IssueService.java
@@ -32,7 +32,7 @@ public interface IssueService {
     Call<Issue> getIssue(
             @Path("owner") String owner,
             @Path("repo") String repo,
-            @Path("number") int number);
+            @Path("number") long number);
 
     @Headers("Accept: application/vnd.github.mockingbird-preview")
     @GET("repos/{owner}/{repo}/issues/{issue_number}/timeline")
